stockwise/requirements.txt
flask
python-dotenv
requests
google-generativeai
yfinance
